Agents/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
psutil==7.0.0
pywin32==308
requests==2.32.3
smmap==5.0.2
urllib3==2.3.0
